+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Refactor scope calls to be more DRY.
 - Fix checking if an attribute exists.
 - Create a hook for the controller to mutate search values (e.g. for date formatting) (https://hitower.atlassian.net/browse/WEAP-187).
+- Print any invalid relation names in the exception.
 
 ## v0.0.15
 - Return only the pagination keys if the request is paginated.